Decouple unmarshaling from receiver and export bulk unmarshaling
To support enhanced versions of fabtoken where we have issues / transfers co-existing with
other operations, one might want to implement a driver that classifies whether a token request
is a vanilla token request or carries enhanced operations.

In case it is a vanilla token request, the enhanced driver may route the token request to
the existing vanilla fabtoken driver, else to handle the advanced operations.

For that to work, I am exporting the unmarshaling functions of the fabtoken driver in order
to facilitate cross driver invocation.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>