Add token request extractor to TCC
Refactor TCC that the tokenRequest extract logic can be injected. This
allows to use alternativ implmentations of the extracting logic other
than the default transient-based logic. A use case for this is, the
FPC-TCC. Currently, FPC does not support transient data, so the
tokenRequest must be provided via function args.

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>