use time.NewTimer instead of time.After
Using time.After enqueues an entry in Go's runtime scheduler and doesn't garbage collect it even if the function the timer was fired from terminates.
As a result, there can be a temporary high memory pressure and it degrades performance.

Replaced all places where time.After was used except in places where we use it for timing periodical processes.

Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>